requirements.txt
jupyter notebook
psycopg2-binary==2.8.6
numpy==1.20.3
pandas==1.2.4
stockstats==0.3.2
tapy==1.9.1
matplotlib==3.4.2
pymysql=1.0.2
workers==0.1
